Update L_Main.umap
Added island in, a plane to be water, rocks around the edge to stop players being able to leave the area and set some placeholders for where the Sunseekers should be as well as a placeholder for the players parking position on the atoll dock. none of the placeholders are visible during play and also have collisions set to no-collision.