created player controller for boat, created game mode and updated player BP
created c++ player controller, created new game mode for asset test and set the custom player controller and the player BP to be the spawned actors, also set up camera and spring arm to boat and adjusted it slightly to a better position. Assigned new game mode to be the chosen game mode in project settings maps and modes.